fix: assemble ts libdefs into single file
closes #386
